+++ persistence/relational-jdbc/src/main/resources/cockroachdb/schema-v4.sql
@@ -28,8 +28,8 @@
 --    policy_mapping_record, events, idempotency_records, scan_metrics_report, commit_metrics_report
 --  * Compatible with PostgreSQL wire protocol
 
-CREATE SCHEMA IF NOT EXISTS POLARIS_SCHEMA;
-SET search_path TO POLARIS_SCHEMA;
+CREATE SCHEMA IF NOT EXISTS ${schema};
+SET search_path TO ${schema};
 
 CREATE TABLE IF NOT EXISTS version (
     version_key TEXT PRIMARY KEY,