+++ persistence/relational-jdbc/src/main/resources/h2/schema-v3.sql
@@ -20,8 +20,8 @@
 -- Changes from v2:
 --  * Added `events` table
 
-CREATE SCHEMA IF NOT EXISTS POLARIS_SCHEMA;
-SET SCHEMA POLARIS_SCHEMA;
+CREATE SCHEMA IF NOT EXISTS ${schema};
+SET SCHEMA ${schema};
 
 CREATE TABLE IF NOT EXISTS version (
     version_key VARCHAR PRIMARY KEY,