+++ persistence/relational-jdbc/src/main/resources/postgres/schema-v4.sql
@@ -20,8 +20,8 @@
 --  * Added `events` table
 --  * Added `idempotency_records` table for REST idempotency
 
-CREATE SCHEMA IF NOT EXISTS POLARIS_SCHEMA;
-SET search_path TO POLARIS_SCHEMA;
+CREATE SCHEMA IF NOT EXISTS ${schema};
+SET search_path TO ${schema};
 
 CREATE TABLE IF NOT EXISTS version (
     version_key TEXT PRIMARY KEY,