+++ persistence/relational-jdbc/src/test/java/org/apache/polaris/persistence/relational/jdbc/MetricsReportPersistenceTest.java
@@ -189,5 +189,10 @@ class MetricsReportPersistenceTest {
     public Optional<String> databaseType() {
       return Optional.empty();
     }
+
+    @Override
+    public Optional<String> schemaName() {
+      return Optional.empty();
+    }
   }
 }