+++ persistence/relational-jdbc/src/test/java/org/apache/polaris/persistence/relational/jdbc/SimpleRelationalJdbcConfiguration.java
@@ -56,4 +56,9 @@ final class SimpleRelationalJdbcConfiguration implements RelationalJdbcConfigura
   public Optional<String> databaseType() {
     return Optional.of(databaseType);
   }
+
+  @Override
+  public Optional<String> schemaName() {
+    return Optional.empty();
+  }
 }